move attribute form and fix mass action
