+++ packages/Webkul/Admin/src/Http/Controllers/Contact/Persons/PersonController.php
@@ -12,7 +12,7 @@ use Webkul\Admin\DataGrids\Contact\PersonDataGrid;
 use Webkul\Admin\Http\Controllers\Controller;
 use Webkul\Admin\Http\Requests\MassDestroyRequest;
 use Webkul\Admin\Http\Resources\PersonResource;
-use Webkul\Attribute\Http\Requests\AttributeForm;
+use Webkul\Admin\Http\Requests\AttributeForm;
 use Webkul\Contact\Repositories\PersonRepository;
 
 class PersonController extends Controller
