+++ packages/Webkul/Admin/src/Http/Controllers/Products/ProductController.php
@@ -11,7 +11,7 @@ use Webkul\Admin\DataGrids\Product\ProductDataGrid;
 use Webkul\Admin\Http\Controllers\Controller;
 use Webkul\Admin\Http\Requests\MassDestroyRequest;
 use Webkul\Admin\Http\Resources\ProductResource;
-use Webkul\Attribute\Http\Requests\AttributeForm;
+use Webkul\Admin\Http\Requests\AttributeForm;
 use Webkul\Product\Repositories\ProductRepository;
 
 class ProductController extends Controller
